+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   - #516 : Dispatch handle : dispatching to an expression 
   - #517 : Interface dispatcher : Check if all abstract methods are implemented 
   - #520 : Parent multi-method default handler is not invoked by its children 
+  - #526 : dispatch operator should work with expression as well 
 - IDE
   - several usability bugs fixed
 - TOOLS